refactor to clean up
